pyproject.toml
[project]
name = "gostpdf"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = ["pyyaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
